# pkg/guacrest/server/server.go
//
// Copyright 2024 The GUAC Authors.
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//     http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

package server

import (
	"context"
	"fmt"
	"net/http"
	"net/url"
	"sort"
	"strings"
	"time"

	"github.com/Khan/genqlient/graphql"
	"github.com/guacsec/guac/pkg/dependencies"
	gen "github.com/guacsec/guac/pkg/guacrest/generated"
	"github.com/guacsec/guac/pkg/guacrest/pagination"
	"github.com/guacsec/guac/pkg/logging"
)

// DefaultServer implements the API, backed by the GraphQL Server
type DefaultServer struct {
	gqlClient graphql.Client
}

func NewDefaultServer(gqlClient graphql.Client) *DefaultServer {
	return &DefaultServer{gqlClient: gqlClient}
}

// Adds the logger to the http request context
func AddLoggerToCtxMiddleware(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		newCtx := logging.WithLogger(r.Context())
		newReq := r.WithContext(newCtx)
		next.ServeHTTP(w, newReq)
	})
}

// Logs data for a request and its response. The request context should already contain
// the logger.
func LogRequestsMiddleware(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		start := time.Now()
		newCtx := logging.WithLogger(r.Context(),
			"method", r.Method,
			"path", r.URL.Path,
		)
		newReq := r.WithContext(newCtx)
		next.ServeHTTP(w, newReq)

		logger := logging.FromContext(newReq.Context())
		logger.Infow("Request handled successfully", "latency", time.Since(start))
	})
}

func (s *DefaultServer) HealthCheck(ctx context.Context, request gen.HealthCheckRequestObject) (gen.HealthCheckResponseObject, error) {
	return gen.HealthCheck200JSONResponse("Server is healthy"), nil
}

func (s *DefaultServer) AnalyzeDependencies(ctx context.Context, request gen.AnalyzeDependenciesRequestObject) (gen.AnalyzeDependenciesResponseObject, error) {
	switch request.Params.Sort {
	case gen.Frequency:
		packages, err := dependencies.GetDependenciesBySortedDependentCnt(ctx, s.gqlClient)
		if err != nil {
			return gen.AnalyzeDependencies500JSONResponse{
				InternalServerErrorJSONResponse: gen.InternalServerErrorJSONResponse{
					Message: err.Error(),
				},
			}, nil
		}

		packageNames := make([]gen.PackageName, 0, len(packages))
		for _, p := range packages {
			packageNames = append(packageNames, gen.PackageName{
				Name:           p.Name,
				DependentCount: p.DependentCount,
			})
		}
		// Stable secondary sort by Name so ties on DependentCount are deterministic
		// across calls — required for pagination cursors to be valid.
		sort.SliceStable(packageNames, func(i, j int) bool {
			if packageNames[i].DependentCount != packageNames[j].DependentCount {
				return packageNames[i].DependentCount > packageNames[j].DependentCount
			}
			return packageNames[i].Name < packageNames[j].Name
		})

		page, info, err := pagination.Paginate(ctx, packageNames, request.Params.PaginationSpec)
		if err != nil {
			return gen.AnalyzeDependencies400JSONResponse{
				BadRequestJSONResponse: gen.BadRequestJSONResponse{Message: err.Error()},
			}, nil
		}

		return gen.AnalyzeDependencies200JSONResponse{
			PackageNameListJSONResponse: gen.PackageNameListJSONResponse{
				PaginationInfo:  info,
				PackageNameList: page,
			},
		}, nil
	case gen.Scorecard:
		return nil, fmt.Errorf("scorecard sort is unimplemented")
	default:
		return nil, fmt.Errorf("%v sort is unsupported", request.Params.Sort)
	}
}

func (s *DefaultServer) GetPackagePurls(ctx context.Context, request gen.GetPackagePurlsRequestObject) (gen.GetPackagePurlsResponseObject, error) {
	unescapedPurl, err := url.PathUnescape(request.Purl)
	if err != nil {
		return gen.GetPackagePurls400JSONResponse{
			BadRequestJSONResponse: gen.BadRequestJSONResponse{
				Message: fmt.Sprintf("failed to unescape package url: %v", err),
			},
		}, nil
	}

	purls, err := FindMatchingPurls(ctx, s.gqlClient, unescapedPurl)
	if err != nil {
		errResp, ok := handleErr(ctx, err, GetPackagePurls).(gen.GetPackagePurlsResponseObject)
		if ok {
			return errResp, nil
		}
		return gen.GetPackagePurls400JSONResponse{
			BadRequestJSONResponse: gen.BadRequestJSONResponse{
				Message: err.Error(),
			},
		}, nil
	}

	sort.Strings(purls)
	page, info, err := pagination.Paginate(ctx, purls, request.Params.PaginationSpec)
	if err != nil {
		return gen.GetPackagePurls400JSONResponse{
			BadRequestJSONResponse: gen.BadRequestJSONResponse{Message: err.Error()},
		}, nil
	}

	return gen.GetPackagePurls200JSONResponse{
		PurlListJSONResponse: gen.PurlListJSONResponse{
			PaginationInfo: info,
			PurlList:       page,
		},
	}, nil
}

func (s *DefaultServer) GetPackageVulns(ctx context.Context, request gen.GetPackageVulnsRequestObject) (gen.GetPackageVulnsResponseObject, error) {
	unescapedPurl, err := url.PathUnescape(request.Purl)
	if err != nil {
		return gen.GetPackageVulns400JSONResponse{
			BadRequestJSONResponse: gen.BadRequestJSONResponse{
				Message: fmt.Sprintf("failed to unescape package url: %v", err),
			},
		}, nil
	}

	includeDeps := false
	if request.Params.IncludeDependencies != nil {
		includeDeps = *request.Params.IncludeDependencies
	}

	vulns, err := GetVulnsForPackage(ctx, s.gqlClient, unescapedPurl, includeDeps)
	if err != nil {
		errResp, ok := handleErr(ctx, err, GetPackageVulns).(gen.GetPackageVulnsResponseObject)
		if ok {
			return errResp, nil
		}
		return gen.GetPackageVulns400JSONResponse{
			BadRequestJSONResponse: gen.BadRequestJSONResponse{
				Message: err.Error(),
			},
		}, nil
	}

	sortVulns(vulns)
	page, info, err := pagination.Paginate(ctx, vulns, request.Params.PaginationSpec)
	if err != nil {
		return gen.GetPackageVulns400JSONResponse{
			BadRequestJSONResponse: gen.BadRequestJSONResponse{Message: err.Error()},
		}, nil
	}

	return gen.GetPackageVulns200JSONResponse{
		VulnerabilityListJSONResponse: gen.VulnerabilityListJSONResponse{
			PaginationInfo:    info,
			VulnerabilityList: page,
		},
	}, nil
}

func (s *DefaultServer) GetPackageDeps(ctx context.Context, request gen.GetPackageDepsRequestObject) (gen.GetPackageDepsResponseObject, error) {
	unescapedPurl, err := url.PathUnescape(request.Purl)
	if err != nil {
		return nil, fmt.Errorf("failed to unescape package url: %w", err)
	}

	purls, err := GetDepsForPackage(ctx, s.gqlClient, unescapedPurl)
	if err != nil {
		errResp, ok := handleErr(ctx, err, GetPackageDeps).(gen.GetPackageDepsResponseObject)
		if ok {
			return errResp, nil
		} else {
			return gen.GetPackageDeps400JSONResponse{
				BadRequestJSONResponse: gen.BadRequestJSONResponse{
					Message: err.Error(),
				},
			}, nil
		}
	}

	depPurls := make([]string, 0, len(purls))
	for _, depPurl := range purls {
		depPurls = append(depPurls, depPurl)
	}
	sort.Strings(depPurls)

	page, info, err := pagination.Paginate(ctx, depPurls, request.Params.PaginationSpec)
	if err != nil {
		return gen.GetPackageDeps400JSONResponse{
			BadRequestJSONResponse: gen.BadRequestJSONResponse{Message: err.Error()},
		}, nil
	}

	return gen.GetPackageDeps200JSONResponse{
		PurlListJSONResponse: gen.PurlListJSONResponse{
			PaginationInfo: info,
			PurlList:       page,
		},
	}, nil
}

func (s *DefaultServer) GetArtifactVulns(ctx context.Context, request gen.GetArtifactVulnsRequestObject) (gen.GetArtifactVulnsResponseObject, error) {
	unescapedDigest, err := url.PathUnescape(request.Digest)
	if err != nil {
		return gen.GetArtifactVulns400JSONResponse{
			BadRequestJSONResponse: gen.BadRequestJSONResponse{
				Message: fmt.Sprintf("failed to unescape digest: %v", err),
			},
		}, nil
	}

	vulns, err := GetVulnsForArtifact(ctx, s.gqlClient, unescapedDigest)
	if err != nil {
		errResp, ok := handleErr(ctx, err, GetArtifactVulns).(gen.GetArtifactVulnsResponseObject)
		if ok {
			return errResp, nil
		}
		return gen.GetArtifactVulns400JSONResponse{
			BadRequestJSONResponse: gen.BadRequestJSONResponse{
				Message: err.Error(),
			},
		}, nil
	}

	sortVulns(vulns)
	page, info, err := pagination.Paginate(ctx, vulns, request.Params.PaginationSpec)
	if err != nil {
		return gen.GetArtifactVulns400JSONResponse{
			BadRequestJSONResponse: gen.BadRequestJSONResponse{Message: err.Error()},
		}, nil
	}

	return gen.GetArtifactVulns200JSONResponse{
		VulnerabilityListJSONResponse: gen.VulnerabilityListJSONResponse{
			PaginationInfo:    info,
			VulnerabilityList: page,
		},
	}, nil
}

func (s *DefaultServer) GetArtifactDeps(ctx context.Context, request gen.GetArtifactDepsRequestObject) (gen.GetArtifactDepsResponseObject, error) {
	unescapedDigest, err := url.PathUnescape(request.Digest)
	if err != nil {
		return nil, fmt.Errorf("failed to unescape digest: %w", err)
	}

	purls, err := GetDepsForArtifact(ctx, s.gqlClient, unescapedDigest)
	if err != nil {
		errResp, ok := handleErr(ctx, err, GetArtifactDeps).(gen.GetArtifactDepsResponseObject)
		if ok {
			return errResp, nil
		} else {
			return gen.GetArtifactDeps400JSONResponse{
				BadRequestJSONResponse: gen.BadRequestJSONResponse{
					Message: err.Error(),
				},
			}, nil
		}
	}

	depPurls := make([]string, 0, len(purls))
	for _, depPurl := range purls {
		depPurls = append(depPurls, depPurl)
	}
	sort.Strings(depPurls)

	page, info, err := pagination.Paginate(ctx, depPurls, request.Params.PaginationSpec)
	if err != nil {
		return gen.GetArtifactDeps400JSONResponse{
			BadRequestJSONResponse: gen.BadRequestJSONResponse{Message: err.Error()},
		}, nil
	}

	return gen.GetArtifactDeps200JSONResponse{
		PurlListJSONResponse: gen.PurlListJSONResponse{
			PaginationInfo: info,
			PurlList:       page,
		},
	}, nil
}

// sortVulns orders vulnerabilities deterministically so that pagination cursors
// are stable across calls. The order is by purl, then by the joined vulnerability
// IDs as a tiebreaker.
func sortVulns(vulns []gen.Vulnerability) {
	sort.Slice(vulns, func(i, j int) bool {
		if vulns[i].Package != vulns[j].Package {
			return vulns[i].Package < vulns[j].Package
		}
		return strings.Join(vulns[i].Vulnerability.VulnerabilityIDs, ",") <
			strings.Join(vulns[j].Vulnerability.VulnerabilityIDs, ",")
	})
}